// src/dto/api/FriendsGetRequest.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface FriendsGetRequest {
  /**
   * If set to true, you will also get back the info from the 'friendof' mode. Some clients show friends and friendof data in separate tabs/panes. If you are always going to load both, use this flag (as opposed to a tabbed dialog approach, where the user may not go to the second tab and thus would not need to load the friendof data.) friendof request variables can be used.
   */
  include_friend_of?: boolean;
  /**
   * If set to 1, you will also get back the info from the 'getfriendgroups' mode. See above for the reason why this would be useful.
   */
  include_groups?: boolean;
  /**
   * If set to a numeric value greater than zero, this mode will only return the number of results indicated. Useful only for building pretty lists for display which might have a button to view the full list nearby.
   */
  friend_limit?: number;
  /**
   * If set to true, you will also get back the birthdays with the friends results below.
   */
  include_bdays?: number;
  [k: string]: any;
}