// src/dto/api/FriendsGetResponse.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface FriendsGetResponse {
  /**
   * List of available friend groups, if include_groups was in the request
   */
  friend_groups?: FriendGroup[];
  /**
   * List of users that list this user as a friend, if include_friend_of was in the request
   */
  friend_ofs?: Friend[];
  /**
   * List of users that this user lists as a friend
   */
  friends: Friend[];
  [k: string]: any;
}
export interface FriendGroup {
  /**
   * The bit number of this friend group, from 1-30
   */
  id: number;
  /**
   * Name of this friend group
   */
  name: string;
  /**
   * The sort order of this friend group, from 0-255
   */
  sortorder: number;
  /**
   * If this friend group is public
   */
  public: boolean;
  [k: string]: any;
}
export interface Friend {
  /**
   * The username of the friend
   */
  username: string;
  /**
   * The full name of the friend
   */
  fullname: string;
  /**
   * If type is identity: Type of identity user, OpenID, TypeKey, etc
   */
  identity_type?: string;
  /**
   * If type is identity: Value for identity, OpenID is a URL
   */
  identity_value?: string;
  /**
   * If type is identity: Pretty display name of an identity user
   */
  identity_display?: string;
  /**
   * The type of journal the friend item is. This value can be one of 'community', 'syndicated', 'news', 'shared', or identity (I), depending on the journaltype of the account in question. The account is a normal personal account when this value is not sent
   */
  type?: string;
  /**
   * The birthday is returned only with the getfriends mode, and only if the user has set their birthday to be publicly visible on their profile. To receive this field you will also need to request it, using the includebdays key
   */
  birthday?: string;
  /**
   * The foreground color of the friend item
   */
  fgcolor: string;
  /**
   * The background color of the friend item
   */
  bgcolor: string;
  /**
   * If the group mask is not 1 (just bit 0 set), then this variable is returned with an 32-bit unsigned integer with a bit 0 always set, and bits 1-30 set for each group this friend is a part of. Bit 31 is reserved.
   */
  group_mask?: number;
  [k: string]: any;
}